arrumei a organização de pastas criando um package para cada classe mudei o nome da classe automationRunner para AdicionaCarrinhoRunner
