--- src/test/java/automationTest/acceptance/AutomationTestRunner.java
@@ -1,13 +1,0 @@
-package automationTest.acceptance;
-
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        features = "src/test/resources/features/ValidaTelaCarrinho.feature",
-        glue = {"automationTest.acceptance.steps.telaCarrinho"}
-)
-public class AutomationTestRunner {
-}
